There is now a map that an @ moves around on.
Credits to that pastebin floating around that had code I used to learn
from. Things of note: mvaddch takes and outputs characters while mvprintw
(the one I used before) invokes printf so it takes strings. Also that q
will quit since the while loop basically sets ch to getch() and then runs
as long as ch is not equal to q.

I need to take closer look at the yy and xx part that draws the map. I
believe what it does is that it keeps redrawing the map over and over
again. Will comment on it when I grok it more.